pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.53.0",
    "fastembed>=0.7.3",
    "google-genai>=1.38.0",
    "llama-index>=0.14.2",
    "llama-index-embeddings-fastembed>=0.5.0",
    "llama-index-node-parser-docling>=0.4.1",
    "llama-index-readers-docling>=0.4.1",
    "llama-index-vector-stores-qdrant>=0.8.5",
    "llmsherpa>=0.1.4",
    "qdrant-client>=1.15.1",
    "ragas>=0.3.5",
    "tesserocr>=2.8.0",
    "uvicorn>=0.36.0",
]
